rest api done

+++ controller/feeds.js
@@ -2,6 +2,7 @@ const { validationResult }=require('express-validator');
 const Post=require('../models/posts');
 const path=require('path')
 const fs=require('fs');
+const User=require('../models/user');
 
 let totalItems;
 exports.getPosts=(req,res,next)=>{
@@ -28,6 +29,7 @@ exports.getPosts=(req,res,next)=>{
 };
 
 exports.createPosts=(req,res,next)=>{
+    let creator;
     const title=req.body.title;
     const content=req.body.title;
     const errors=validationResult(req);
@@ -46,14 +48,24 @@ exports.createPosts=(req,res,next)=>{
     const post=new Post({
         title:title,
         content:content,
-        creator:{name:'Rupam'},
+        creator:req.userId,
         imageUrl:imageUrl,
+
     })
     post.save()
-    .then(postReturned=>{
-        res.status(200).json({
+    .then(result=>{
+        return User.findById(req.userId);
+    })
+    .then(user=>{
+        creator=user;
+        user.posts.push(post);
+        return user.save();
+    })
+    .then(result=>{
+        res.status(201).json({
             message:"Post created successfully",
-            post:postReturned
+            post:post,
+            creator:{_id:creator._id,name:creator.name}
         })
     })
     .catch(err=>{
@@ -112,6 +124,13 @@ exports.updatePost=(req,res,next)=>{
             error.statusCode=404;
             throw error;
         }
+        if(!(post.creator.toString()===req.userId)){
+            console.log('post==='+post)
+            console.log(req.userId)
+            const error=new Error('could not update post');
+            error.statusCode=404;
+            throw error;
+        }
         if(imageUrl!=post.imageUrl){
             clearImage(post.imageUrl);
         }
@@ -140,11 +159,24 @@ exports.deletePost=(req,res,next)=>{
                 error.statusCode=404;
                 throw error;
             }
-
+            if(!(post.creator.toString()===req.userId)){
+                console.log('post==='+post)
+                console.log(req.userId)
+                const error=new Error('could not update post');
+                error.statusCode=404;
+                throw error;
+            }
             clearImage(post.imageUrl);
             return Post.findByIdAndRemove(postId);
         })
-        .then(post=>{
+        .then(resultt=>{
+            return User.findById(req.userId)
+        })
+        .then(user=>{
+            user.posts.pull(postId);
+            return user.save();
+        })
+        .then(afterSave=>{
                 res.status(200).json({message:'Deleted Post'})
         })
         .catch(err=>{
